% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{flowSorted}
\alias{flowSorted}
\title{A subset of FlowSorted.DLPFC.450k data set}
\format{A MethylSet object with 1e4 rows (probes)
and 58 columns (samples).}
\description{
This is the script used to create a subset of the
FlowSorted.DLPFC.450k data set from Bioconductor.
The purpose is to create an example object for the man
pages and vignette in this package.

The object was created using the script in /inst and
located in the /data folder.
}

