% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{MSbetween}
\alias{MSbetween}
\alias{MSbetween,quantro-method}
\alias{MSbetween.quantro}
\title{Accessors for the 'MSbetween' slot of a quantro object.}
\usage{
MSbetween(object)

\S4method{MSbetween}{quantro}(object)

\S4method{MSbetween}{quantro}(object)
}
\arguments{
\item{object}{a \code{quantro} object}
}
\description{
Accessors for the 'MSbetween' slot of a quantro object.
}
\examples{
library(minfi)
data(flowSorted)
pd <- pData(flowSorted)
qtest <- quantro(flowSorted, groupFactor = pd$CellType)
MSbetween(qtest)
}

