% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{MSwithin}
\alias{MSwithin}
\alias{MSwithin,quantro-method}
\alias{MSwithin.quantro}
\title{Accessors for the 'MSwithin' slot of a quantro object.}
\usage{
MSwithin(object)

\S4method{MSwithin}{quantro}(object)

\S4method{MSwithin}{quantro}(object)
}
\arguments{
\item{object}{a \code{quantro} object}
}
\description{
Accessors for the 'MSwithin' slot of a quantro object.
}
\examples{
library(minfi)
data(flowSorted)
pd <- pData(flowSorted)
qtest <- quantro(flowSorted, groupFactor = pd$CellType)
MSwithin(qtest)
}

