% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{quantroPvalPerm}
\alias{quantroPvalPerm}
\alias{quantroPvalPerm,quantro-method}
\alias{quantroPvalPerm.quantro}
\title{Accessors for the 'quantroPvalPerm' slot of a quantro object.}
\usage{
quantroPvalPerm(object)

\S4method{quantroPvalPerm}{quantro}(object)

\S4method{quantroPvalPerm}{quantro}(object)
}
\arguments{
\item{object}{a \code{quantro} object}
}
\description{
Accessors for the 'quantroPvalPerm' slot of a quantro object.
}
\examples{
library(minfi)
data(flowSorted)
pd <- pData(flowSorted)
qtest <- quantro(flowSorted, groupFactor = pd$CellType)
quantroPvalPerm(qtest)
}

