% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,quantro-method}
\alias{summary.quantro}
\title{Accessors for the 'summary' slot of a quantro object.}
\usage{
summary(object, ...)

\S4method{summary}{quantro}(object)

\S4method{summary}{quantro}(object)
}
\arguments{
\item{object}{a \code{quantro} object}

\item{...}{other}
}
\description{
Accessors for the 'summary' slot of a quantro object.
}
\examples{
library(minfi)
data(flowSorted)
pd <- pData(flowSorted)
qtest <- quantro(flowSorted, groupFactor = pd$CellType)
summary(qtest)
}

