% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{anova}
\alias{anova}
\alias{anova,quantro-method}
\alias{anova.quantro}
\title{Accessors for the 'anova' slot of a quantro object.}
\usage{
anova(object, ...)

\S4method{anova}{quantro}(object)

\S4method{anova}{quantro}(object)
}
\arguments{
\item{object}{a \code{quantro} object}

\item{...}{additional objects of the same type.}
}
\description{
Accessors for the 'anova' slot of a quantro object.
}
\examples{
library(minfi)
data(flowSorted)
pd <- pData(flowSorted)
qtest <- quantro(flowSorted, groupFactor = pd$CellType)
anova(qtest)
}

