% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{anova}
\alias{anova}
\alias{anova,quantro-method}
\alias{anova.quantro}
\title{Accessors for the 'anova' slot of a quantro object.}
\usage{
anova(object, ...)

\S4method{anova}{quantro}(object)

\S4method{anova}{quantro}(object)
}
\arguments{
\item{object}{a \code{quantro} object}
}
\description{
Accessors for the 'anova' slot of a quantro object.
}
\examples{
library(minfi)
data(flowSorted)
pd <- pData(flowSorted)
qtest <- quantro(flowSorted, groupFactor = pd$CellType)
anova(qtest)
}

