% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{quantroStatPerm}
\alias{quantroStatPerm}
\alias{quantroStatPerm,quantro-method}
\alias{quantroStatPerm.quantro}
\title{Accessors for the 'quantroStatPerm' slot of a quantro object.}
\usage{
quantroStatPerm(object)

\S4method{quantroStatPerm}{quantro}(object)

\S4method{quantroStatPerm}{quantro}(object)
}
\arguments{
\item{object}{a \code{quantro} object}
}
\description{
Accessors for the 'quantroStatPerm' slot of a quantro object.
}
\examples{
library(minfi)
data(flowSorted)
pd <- pData(flowSorted)
qtest <- quantro(flowSorted, groupFactor = pd$CellType)
quantroStatPerm(qtest)
}

